Rename fil tilbake til gammelt navn
